[#STATISTICS]: get all data

+++ src/Statistics/Infrastructure/database/migrations/2024_07_22_194942_create_monthly_statistics_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('user_id')->index();
             $table->unsignedSmallInteger('year');
             $table->enum('month', MonthEnum::values());
-            $table->float('total_expense')->default(0);
-            $table->float('total_income')->default(0);
+            $table->decimal('total_expense', 15)->default(0);
+            $table->decimal('total_income', 15)->default(0);
             $table->timestamps();
         });
     }
